+++ packages/react-sdk/src/core/components/CallLayout/PaginatedGridLayout.tsx
@@ -11,7 +11,7 @@ import {
 import { ParticipantsAudio } from '../Audio';
 import { IconButton } from '../../../components';
 import { chunk } from '../../../utilities';
-import { usePaginatedLayoutSortPreset } from './hooks';
+import { useFilteredParticipants, usePaginatedLayoutSortPreset } from './hooks';
 
 const GROUP_SIZE = 16;
 
@@ -71,6 +71,11 @@ export type PaginatedGridLayoutProps = {
   excludeLocalParticipant?: boolean;
 
   /**
+   * Predicate to filter call participants.
+   */
+  filterParticipants?: (participant: StreamVideoParticipant) => boolean;
+
+  /**
    * When set to `false` disables mirroring of the local partipant's video.
    * @default true
    */
@@ -89,6 +94,7 @@ export const PaginatedGridLayout = (props: PaginatedGridLayoutProps) => {
       ? props.groupSize || GROUP_SIZE
       : GROUP_SIZE,
     excludeLocalParticipant = false,
+    filterParticipants,
     mirrorLocalParticipantVideo = true,
     pageArrowsVisible = true,
     VideoPlaceholder,
@@ -101,9 +107,12 @@ export const PaginatedGridLayout = (props: PaginatedGridLayoutProps) => {
   ] = useState<HTMLDivElement | null>(null);
 
   const call = useCall();
-  const { useParticipants, useRemoteParticipants } = useCallStateHooks();
-  const participants = useParticipants();
+  const { useRemoteParticipants } = useCallStateHooks();
   const remoteParticipants = useRemoteParticipants();
+  const participants = useFilteredParticipants({
+    excludeLocalParticipant,
+    filterParticipants,
+  });
 
   usePaginatedLayoutSortPreset(call);
 
@@ -117,12 +126,8 @@ export const PaginatedGridLayout = (props: PaginatedGridLayoutProps) => {
 
   // only used to render video elements
   const participantGroups = useMemo(
-    () =>
-      chunk(
-        excludeLocalParticipant ? remoteParticipants : participants,
-        groupSize,
-      ),
-    [excludeLocalParticipant, remoteParticipants, participants, groupSize],
+    () => chunk(participants, groupSize),
+    [participants, groupSize],
   );
 
   const pageCount = participantGroups.length;